+++ app/src/main/java/aandroid/paandroidportfolio/tiptracker/trip/Trip.kt
@@ -6,10 +6,10 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "TripDatabase")
 data class Trip(
-    val money: Int,
-    var mileage: Int,
+    val money: Float,
+    var mileage: Float,
     var date: String,
-    var hours: Int,
+    var hours: Float,
     var gasprice: Float,
     @PrimaryKey(autoGenerate = true)
     val id: Int?=null
